React week 2 HW added.
