+++ src/client/components/MealsComponent/Meal.css
@@ -1,0 +1,10 @@
+.meal-card {
+  border: 1px solid #ccc;
+  padding: 10px;
+  margin: 10px;
+}
+
+.meal-image {
+  width: 100%;
+  height: auto;
+}
