--- src/client/components/MealsComponent/MealsList.css
@@ -1,3 +1,0 @@
-section.meals-container {
-  color: green;
-}
